Add files via upload
In this project, we will show how to use the Java Swing library to build a digital clock. This digital clock uses a graphical user interface(GUI) to display current date and time. This clock is build using the Java Swing library, which provides a flexible set of components for creating a graphical user interface(GUI). This application has an easy and simple design with Labels to show current date and time. Grab a cup of coffee, sit down and start using the Digital Clock.